+++ src/distribution/chi_squared.rs
@@ -1,5 +1,5 @@
 use Result;
-use distribution::{Continuous, Distribution, Gamma, Univariate};
+use distribution::{Continuous, Distribution, Gamma, Univariate, WeakRngDistribution};
 use rand::Rng;
 use rand::distributions::{IndependentSample, Sample};
 use statistics::*;
@@ -142,6 +142,8 @@ impl Distribution<f64> for ChiSquared {
         self.g.sample(r)
     }
 }
+
+impl WeakRngDistribution<f64> for ChiSquared {}
 
 impl Univariate<f64, f64> for ChiSquared {
     /// Calculates the cumulative distribution function for the