+++ src/distribution/inverse_gamma.rs
@@ -1,5 +1,5 @@
 use {Result, StatsError};
-use distribution::{Continuous, Distribution, Univariate};
+use distribution::{Continuous, Distribution, Univariate, WeakRngDistribution};
 use function::gamma;
 use rand::Rng;
 use rand::distributions::{IndependentSample, Sample};
@@ -131,6 +131,8 @@ impl Distribution<f64> for InverseGamma {
         1.0 / super::gamma::sample_unchecked(r, self.shape, self.rate)
     }
 }
+
+impl WeakRngDistribution<f64> for InverseGamma {}
 
 impl Univariate<f64, f64> for InverseGamma {
     /// Calculates the cumulative distribution function for the inverse gamma