+++ src/distribution/weibull.rs
@@ -1,5 +1,5 @@
 use {Result, StatsError, consts};
-use distribution::{Continuous, Distribution, Univariate};
+use distribution::{Continuous, Distribution, Univariate, WeakRngDistribution};
 use function::gamma;
 use rand::Rng;
 use rand::distributions::{IndependentSample, Sample};
@@ -133,6 +133,8 @@ impl Distribution<f64> for Weibull {
         self.scale * (-x.ln()).powf(1.0 / self.shape)
     }
 }
+
+impl WeakRngDistribution<f64> for Weibull {}
 
 impl Univariate<f64, f64> for Weibull {
     /// Calculates the cumulative distribution function for the weibull