+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+HEAD
+ - Create `distribution::WeakRngDistribution` trait for those who don't want to go through the trouble of importing `rand` and deciding on an RNG
+ - Implement `distribution::WeakRngDistribution` for all distributions
+
 v0.9.0
  - implemented infinite sequence generator for periodic sequence
  - implemented infinite sequence generator for sinusoidal sequence